chore: let the tests be in one same file
